+++ ext/himg/src/renderer.rs
@@ -1,7 +1,7 @@
 use crate::html_to_image::html_to_image;
 use crate::options::Options;
 use crate::writer::write_png;
-use crate::logger::{TimedLogger};
+use crate::logger::{Logger, NullLogger, TimedLogger};
 
 pub fn render_blocking(html: String, options: Options) -> Result<Vec<u8>, std::io::Error> {
     let runtime = tokio::runtime::Runtime::new()?;
@@ -11,12 +11,16 @@ pub fn render_blocking(html: String, options: Options) -> Result<Vec<u8>, std::i
 
 // render_to_bytes, render_to_string, render_to_file, render_to_io
 pub async fn render(html: String, options: Options) -> Result<Vec<u8>, std::io::Error> {
-    let mut logger = TimedLogger::init();
+    let mut logger: Box<dyn Logger> = if options.verbose {
+        Box::new(TimedLogger::init())
+    } else {
+        Box::new(NullLogger{})
+    };
 
     // Render to Image
     //let base_url = format!("file://{}", path_string.clone());
     let base_url = None;
-    let render_output = html_to_image(&html, base_url, options, &mut logger).await;
+    let render_output = html_to_image(&html, base_url, options, &mut *logger).await;
 
     // Determine output path, and open a file at that path.
     let mut output_buffer: Vec<u8> = Vec::new();